// types/requests/get-student-classes-enrolled.request.ts
import { EntityStatus } from '../enums/entity-status.enum';

export interface GetStudentClassesEnrolledRequest {
    studentId: string;
    className?: string;
    teacherName?: string;
    schoolName?: string;
    classCode?: string;
    classStatus: EntityStatus;
    schoolId: number;
    pageIndex?: number;
    pageSize?: number;
    sortBy: string;
    sortDirection: string;
    searchTerm?: string;
}